backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
alembic>=1.12.0

# Environment and Settings
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Langchain and AI
langchain>=0.0.300
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5
google-generativeai>=0.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
docx2txt>=0.8
pymupdf>=1.24.0

# Vector DB and Dependencies
chromadb>=0.4.0
numpy>=1.24.0
pandas>=2.0.0

# Additional Dependencies
typing-extensions>=4.8.0
python-dateutil>=2.8.0
requests>=2.31.0
aiofiles>=23.2.0